rename i18next_language_detector should work for both electron and cordova
