+++ app/frontend/platform/i18next_language_detector.js
